services/auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
python-multipart==0.0.12
PyJWT==2.9.0
redis==5.0.8
celery==5.4.0
boto3==1.34.162
pymongo==4.8.0
psycopg[binary]==3.2.1
prometheus-client==0.20.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
